+++ packages/root/src/index.ts
@@ -1,5 +1,8 @@
 export * from "./types";
 
-import Common from "./common";
-import prepNavigate from "./prepNavigate";
-export { Common, prepNavigate };
+import locationUtils from "./locationUtils";
+import keyGenerator from "./keyGenerator";
+import prepareNavigate from "./prepareNavigate";
+import navigationConfirmation from "./navigationConfirmation";
+
+export { locationUtils, keyGenerator, navigationConfirmation, prepareNavigate };