+++ packages/root/types/common.d.ts
@@ -1,2 +1,0 @@
-import { CommonHistory, Options } from "./types/hickory";
-export default function Common<Q>(options?: Options<Q>): CommonHistory<Q>;