+++ packages/root/types/keygen.d.ts
@@ -1,2 +1,0 @@
-import { KeyFns } from "./types/keygen";
-export default function createKeyGenerator(): KeyFns;