+++ packages/root/types/types/navigationConfirmation.d.ts
@@ -1,5 +1,5 @@
 import { SessionLocation } from "./location";
-import { Action } from "./hickory";
+import { Action } from "./navigation";
 export interface NavigationInfo<Q> {
     to: SessionLocation<Q>;
     from: SessionLocation<Q>;
@@ -8,6 +8,10 @@ export interface NavigationInfo<Q> {
 export declare type ConfirmationFunction<Q> = (info: NavigationInfo<Q>, confirm: () => void, prevent?: () => void) => void;
 export interface ConfirmationMethods<Q> {
     confirmNavigation(info: NavigationInfo<Q>, confirm: () => void, prevent?: () => void): void;
+    confirmWith(fn?: ConfirmationFunction<Q>): void;
+    removeConfirmation(): void;
+}
+export interface BlockingHistory<Q> {
     confirmWith(fn?: ConfirmationFunction<Q>): void;
     removeConfirmation(): void;
 }