+++ packages/root/types/types/prepareNavigate.d.ts
@@ -1,7 +1,7 @@
-import { Action, ToArgument, NavType } from "./hickory";
+import { Action, ToArgument, NavType } from "./navigation";
 import { SessionLocation } from "./location";
-import { LocationMethods } from "./locationFactory";
-import { KeyMethods } from "./keygen";
+import { LocationUtils } from "./locationUtils";
+import { KeyFns } from "./keyGenerator";
 export interface PreppedNavigation<Q> {
     action: Action;
     finish(): void;
@@ -9,7 +9,8 @@ export interface PreppedNavigation<Q> {
 }
 export declare type Preparer<Q> = (to: ToArgument<Q>, navType: NavType) => PreppedNavigation<Q>;
 export interface PrepNavigateArgs<Q> {
-    utils: LocationMethods<Q> & KeyMethods;
+    locationUtils: LocationUtils<Q>;
+    keygen: KeyFns;
     current(): SessionLocation<Q>;
     push(l: SessionLocation<Q>): () => void;
     replace(l: SessionLocation<Q>): () => void;